tradeinternal_mcp_server/config.py
"""Configuration helpers for the TradingView MCP server."""

from __future__ import annotations

import os
from dataclasses import dataclass

from dotenv import load_dotenv


@dataclass(frozen=True)
class DatabaseConfig:
    """Container for database connection parameters."""

    host: str
    port: int
    user: str
    password: str
    database: str

    @classmethod
    def from_env(cls) -> "DatabaseConfig":
        """Build a config object by reading the standard environment vars."""

        load_dotenv()

        missing = [var for var in ("DB_HOST", "DB_PORT", "DB_USER", "DB_PASSWORD", "DB_NAME") if not os.getenv(var)]
        if missing:
            missing_vars = ", ".join(missing)
            raise RuntimeError(f"Missing required database environment variables: {missing_vars}")

        return cls(
            host=os.environ["DB_HOST"],
            port=int(os.environ["DB_PORT"]),
            user=os.environ["DB_USER"],
            password=os.environ["DB_PASSWORD"],
            database=os.environ["DB_NAME"],
        )
