requirements.txt
fastmcp>=0.3.0
pymysql>=1.1.0
python-dotenv>=1.0.0
typing_extensions>=4.9.0
fastapi>=0.110.0
uvicorn>=0.23.0
